server/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
praat-parselmouth>=0.4.3
scipy>=1.13.0
numpy>=1.26.0
faster-whisper>=1.0.0
